[lexer] Implementations of block scope characters
- Parentheses handling.
- Square brackets handling.
- Curly braces handling.

[improvements] Check EOF character while using 'while loop'

In the process of tracking literals (INTEGER_LITERAL
for example), the lexer uses 'while loop' to advance
the iterator `currentChar` until it reaches the end
position of the integer. This action checks for EOF
character in case the iterator goes farther than the
end of source file, otherwise the system throws
"Address boundary error".